Move Attr to nodes/
