+++ lib/jsdom/living/attributes.js
@@ -3,8 +3,8 @@ const DOMException = require("../web-idl/DOMException");
 const defineGetter = require("../utils").defineGetter;
 const idlUtils = require("./generated/utils");
 const attrGenerated = require("./generated/Attr");
-const changeAttributeImpl = require("./attributes/Attr-impl").changeAttributeImpl;
-const getAttrImplQualifiedName = require("./attributes/Attr-impl").getAttrImplQualifiedName;
+const changeAttributeImpl = require("./nodes/Attr-impl").changeAttributeImpl;
+const getAttrImplQualifiedName = require("./nodes/Attr-impl").getAttrImplQualifiedName;
 
 // https://dom.spec.whatwg.org/#namednodemap
